+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 - Bump Sentry to v5.7.4  
 - Fix CDSS alert hiding when multiple BPs are added on the same day
 - Bump flipper to v0.145.0
-- [In Progress: 11 May 2022] Show call result status for the appointment in `ContactPatientBottomSheet`
+- Add query to load latest scheduled appointment for the patient
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
+- [In Progress: 11 May 2022] Show call result status for the appointment in `ContactPatientBottomSheet`
+- [In Progress: 18 May 2022] Don't show CDSS alert for patient if appointment is more than 2 days overdue
 
 ### Changes